+++ cmd/vigilante/cmd/reporter.go
@@ -3,7 +3,7 @@ package cmd
 import (
 	"fmt"
 
-	bbnclient "github.com/babylonlabs-io/babylon/v3/client/client"
+	bbnclient "github.com/babylonlabs-io/babylon/v4/client/client"
 	"github.com/spf13/cobra"
 
 	"github.com/babylonlabs-io/vigilante/btcclient"