+++ testutil/datagen/datagen.go
@@ -3,7 +3,7 @@ package datagen
 import (
 	"math/rand"
 
-	"github.com/babylonlabs-io/babylon/v3/testutil/datagen"
+	"github.com/babylonlabs-io/babylon/v4/testutil/datagen"
 	"github.com/babylonlabs-io/vigilante/types"
 )