+++ CHANGELOG.md
@@ -37,6 +37,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 ## Unreleased
 
+* [#455](https://github.com/babylonlabs-io/vigilante/pull/455) chore: ignore warnings from utils package
+
 ### Improvements
 
 * [#434](https://github.com/babylonlabs-io/vigilante/pull/434) chore: avoid adding to pending tracker